==> introduction_to_machine_learning/pandas_micro_course/01_creating_reading_writing.ipynb <==
# To use `Pandas`, you'll typically start with the following line of code.
import pandas as pd

## Creating Data
There are two core objects in pandas: the `DataFrame` and the `Series`

#### DataFrame
A **DataFrame is a table**.\
It contains an array of individual entries, each of which has a certain value.\
Each entry correspoinds to a row (or record) and a column.

pd.DataFrame({'Yes': [50, 21], 'No': [131, 2]})

pd.DataFrame({'Apples': [30], 'Bananas': [21]})

pd.DataFrame({'Bob': ['I like it.', 'It was awful.'], 'Sue': ['Pretty good.', 'Bland.']})

We are using the `pd.DataFrame()` constructor to generate these DataFrame objects.\
The syntax for declaring a new one is a dictionary whose keys are the column names (`Bob` and `Sue` in this example), and\
whose values are a list of entries. This is the standard way of constructing a new DataFrame, and the one you\
are most likely to encounter.

### Row lables
The list of row labels used in a DataFrame is known as an `Index`.\
We can assign values to it by using an `index` parameter in our constructor:

pd.DataFrame({
    'Bob': ['I liked it.', 'It was awful.'],
    'Sue': ['Pretty good.', 'Bland.']},
    index=['Product A', 'Product B'])

### Series
A series, by contrast, is a sequence of data values.\
If a DataFrame is a table, a Series is a list.\
And in fact you can create one with nothing more than al list:

pd.Series([1, 2, 3, 4, 5])

A Series is, in essence, a single column of a DataFrame.\
So you can assign column values to the Series the same way as before, using an index parameter.\
However, a Series does not have a column name, it only has one overall name:

pd.Series([30, 35, 40], index=['2018 Sales', '2019 Sales', '2020 Sales'], name='Product A')

### Reading Data files
Being able to create a DataFrame of Series by hand is handy.\
But, most of the time, we won't actually be creating our own data by hand. Instead, we'll be working with data that already exists.

house_reviews = pd.read_csv('../data/002/melb_data.csv')

house_reviews.shape # check how large the resulting DataFrame
# So our new DataFrame has 13580 records split across 21 different columns.
# That's almost 285000 entries!

We can examine the contents of the resultant DataFrame using the `head()` command, which grabs the first five rows:

house_reviews.head()

# house_reviews = pd.read_csv('../data/002/melb_data.csv', index_col=0)
house_reviews = pd.read_csv('../data/002/melb_data.csv', index_col='Rooms')

house_reviews.head()

### Save this DataFrame to disk as a csv file.

house_reviews.to_csv('ex_01_house_review.csv')
