==> introduction_to_machine_learning/pandas_micro_course/02_indexing_selecting_and_assigning.ipynb <==
import pandas as pd

### Native accessors

house_data = pd.read_csv('../data/002/melb_data.csv')
house_data.head()

### Select 'Address' column from 'house_data' and assign the result to the variable 'reviews'.

reviews = house_data['Address']
reviews

### Select the first value from the 'Address' column of `house_data`, assign it to variable `first_value`

first = house_data.Address[0]
first

first = house_data['Address'][0]
first

### Indexing in pandas
The indexing operator and attribute selection are nice because they work just like they do in the rest of the Python ecosystem.\
As a novice, this makes them easy to pick up and use.\
However, pandas has its own accessor operators, `loc` and `iloc`.\
For more advanced operations, these are the ones you're supposed to be using.

house_data.iloc[:1]

Both `loc` and `iloc` are row-first, column-second. 

This means that it's marginally easier to retrieve rows, and marginally harder to get retrieve columns.\
To get a column with `iloc`, we can do the following:

### `.loc` selects data only by labels
### `.iloc` selects data only by integer location

house_data.iloc[:10]['Address'] # Select the first 10 values from the `Address` column in `house_data`/

house_data.iloc[:10][['Address', 'Price']] # Select the first 10 values from the `Address` and 'Price' columns in `house_data`

On its own, the `:` operator, which also comes from native Python, means "everything".\
When combined with other selectors, however, it can be used to indicate a range of values.

house_data.iloc[1:3, 1]

house_data.iloc[[0,2,3], 1] # output only 1, 3 & 4 strings of 2nd column
# This sounds like 'to pass a list'

house_data.iloc[-5:]

house_data.iloc[0] # Select the first row of data(the first record) from `house_data`

### Label based selection

house_data.loc[[1, 33, 843, 5098, 13575]] # Select the records with index labels 1, 33, 843, 5098, 13575

house_data.loc[1, 'Price'] # To get the second entry in `Price`

`iloc` is conceptually simpler than `loc` because it ignores the dataset's indices.\ 
When we use `iloc` we treat the dataset like a big matrix (a list of lists).\
`loc`, by contrast, uses the information in the indices to do its work.

house_data.loc[:, ['Address', 'Price']] # get the all rows form columns `Address` and `Price`

### Choosing between `loc` and `iloc`
When choosing or tansitioning between `loc` and `iloc`, there is on "gotcha" worth keeping in mind, which is  
that the two methods use slightly different indexing schemes.

`iloc` uses the Python `stdlib` indexing scheme, where the first element of the range is included and the last one excluded.  
So `0:10` will select entries `0, ..., 9`. 

`loc`, meanwhile, indexes inclusively.  
So `0:10` will select entries `0, ..., 10`.

Why the change?  
Remember that `loc` can index any `stdlib` type: `Strings`, for example.  
If we have a DataFrame with index values `Apples, ..., Potatoes, ...`, and we want to select `all the alphabetical fruit choices between Apples and Potatoes`, then it's a lot more convenient to index `df.loc['Apples':'Potatoes']` than it is to index something like `df.loc['Apples', 'Potatoet']`(`t` coming after `s` in the alphabet).

This is particularly confusing when the DataFrame index is a simple numerical list, e.g. `0, ..., 1000`. In this case `df.iloc[0:1000]` will return 1000 entries, while `df.loc[0:1000]`, return 1001 of them! To get 1000 elements using `loc` you will need to go one lower and ask for `df.loc[0:999]`.

Otherwise, the semantics of using `loc` are the same as those for `iloc`.

### Manipulating the index

The `set_index()` method can be used to do the job. Here is what happens when we `set_index` to the `title` field:

house_data.set_index('Price')

This is useful if you can come up with an index for the dataset which is better than the current one.

### Conditional selection

price_filter = house_data.loc[house_data.Price >= 5945000]
price_filter.shape

price_filter.head(6)

bathroom_filter = house_data.loc[house_data.Bathroom == 6]
bathroom_filter.shape

bathroom_filter

mixed_filter = house_data.loc[(house_data.Price >= 3000000) & (house_data.Landsize > 2000)]
mixed_filter.shape

mixed_filter

mixed_filter = house_data.loc[(house_data.Price >= 5000000) | (house_data.Landsize > 4000)]
mixed_filter.shape

mixed_filter.head()

### `isin` and `isnull`

`isin` is lets you select data whose value "is in" a list of values.

is_in = house_data.loc[house_data.Date.isin(["18/03/2017"])]
is_in.head()

`isnull` These methods let you highlight values which are (or are not) empty (NaN).

is_null = house_data.loc[house_data.YearBuilt.isnull()]
is_null.shape

is_null.head()

`notnull`

not_null = house_data.loc[house_data.YearBuilt.notnull()]
not_null.shape

not_null

### Assigning data
Going the other way, assigning data to a DataFrame is easy. You can assign either a constant value:

house_data['Price'] = '$'
house_data.head(3)

**Or with an iterable of values**:

house_data['Address'] = range(len(house_data), 0, -1)
house_data
