==> introduction_to_machine_learning/pandas_micro_course/03_summary_funciton_and_maps.ipynb <==
### Introduction
Plucking the right data out of our data representation is critical to getting work done, as we demonstrated in the exercises.

This tutorial will cover different operations we can apply to our data to get the input "just right".

import pandas as pd

# Get data
wine_catalog = pd.read_csv('../data/wine_catalog/winemag-data_first150k.csv')
wine_catalog.shape

# Find error(not a number values) in row of Dataset
# Since the correct elements are no more than three characters in length, I will find all that more.
x = []
for i, j in enumerate(wine_catalog.points):
    if len(str(j)) > 3:
        x.append(i)
#         print(j) # error's row(item) print 
print(x)    

# Drop(delete) row with errors
for i in x:
    wine_catalog = wine_catalog.drop([i], axis=0)

# Drop(delete) not a number values in 'Points' column of 'Wine Magazine' Dataset.
wine_catalog = wine_catalog.dropna(subset=['points'])

# Change the datatype of one or more columns to numeric
wine_catalog[['points', 'price']] = wine_catalog[['points', 'price']].astype(float)

# Drop(delete) row that over 100 points
x = list(wine_catalog.loc[wine_catalog.points > 100].index)
for i in x:
    wine_catalog = wine_catalog.drop([i], axis=0)

wine_catalog.to_csv('../data/wine_catalog/wine_store_dataset.csv') # create a new catalog(clear catalog)

### Summary functions

Pandas provides many simple "summary functions" (not an official name) which restructure the data in some useful way.  
For example, consider the `describe()` method:

wine_catalog.points.describe()

print(list(wine_catalog.columns)) # get a col. name
wine_catalog.province.describe() # another output winth 'describe()' method

For example, to see the mean of the points allocted (e.g. how well an averagely rated wine does), we can use the `mean()` function.

wine_catalog.points.mean()

To see a list of unique values we can use the `unique()` function

len(wine_catalog.province.unique())

len(sorted(wine_catalog.winery.unique()))

`value_counts` maked a list (Dataset series) of unique values and shows us how often items occur in the Dataset.

wine_catalog.price.value_counts()

# Find value that repeating less than 10 times
winery = []
for i in list(wine_catalog.price.value_counts()):
    if i < 10:
        winery.append(wine_catalog.winery[i])

list(set(winery)) # sort a list

list(wine_catalog['winery'].loc[wine_catalog.price > 1000]) # winery, whose price by a bottle of wine over 1000$

# Sorting values those price great than 1000$
x = wine_catalog.loc[wine_catalog.price >= 1000]
x.sort_values(by=['price'])

### `Series.idxmax` 
Return index of the maximum element.

# (a / b).idxmax() >>> Can be used on values that are only obtained by various real-time operations.
bargain_idx =  (wine_catalog.points / wine_catalog.price).idxmax()
print(bargain_idx)
bargain_wine = wine_catalog.loc[bargain_idx, 'winery']
print(bargain_wine)

wine_catalog.loc[bargain_idx]

### Maps

### `map()`
A `map` is a term, borrowed from mathematics, for a function that takes one set of values and `maps` them to another set of values.

# df.mean(axis = 1, skipna = True) >>> skipna - skip the NaN values while finding the mean
review_points_mean = wine_catalog.points.mean()
review_points_mean

review_points_mean = wine_catalog.points.median()
review_points_mean

wine_catalog.points.map(lambda i: i - review_points_mean) # `i` is an each element in `points` series

The function you pass to `map()` should expect a single value from the Series (a point value, in the above example), and return a transformed version of that value.  
`map()` returns a new Series where all the values have been transformed by your function.

### `apply()`
`apply()` is the equivalent method if we want to transform a whole DataFrame by calling a custom method on each row.

def remean_points(row):
    row.points -= review_points_mean # 94.0 - 87.87346229986456
    return row

wine_catalog.apply(remean_points, axis='columns')

### Finding the same word
Creating a `Series` counting how many times a word appears in `variety` column in the Dataset.

n_word = wine_catalog.variety.map(lambda i: 'Pinot' in i).sum()
descriptor_counts = pd.Series([n_word], index=['Pinot'])
descriptor_counts

wine_catalog.country + '-' + wine_catalog.region_1

### Dataframe or series (if, else and loc)

# We'd like to host these wine reviews on our website,  
# but a rating system ranging from 80 to 100 points is too hard to understand - we'd like to translate them into simple star ratings.  
# A score of 95 or higher counts as 3 stars, a score of at least 85 but less than 95 is 2 stars. Any other score is 1 star.

# Also, the Canadian Vintners Association bought a lot of ads on the site,  
# so any wines from Canada should automatically get 3 stars, regardless of points.

# Create a series star_ratings with the number of stars corresponding to each review in the dataset.

def stars(row):
    if row.country == 'Canada':
        return 4
    elif row.points >= 95:
        return 3
    elif row.points >= 85:
        return 2
    else:
        return 1

star_ratings = wine_catalog.apply(stars, axis='columns')
star_ratings

# We'd like to host these wine reviews on our website,  
# but a rating system ranging from 80 to 100 points is too hard to understand - we'd like to translate them into simple star ratings.  
# A score of 95 or higher counts as 3 stars, a score of at least 85 but less than 95 is 2 stars. Any other score is 1 star.

# Also, the Canadian Vintners Association bought a lot of ads on the site,  
# so any wines from Canada should automatically get 3 stars, regardless of points.

# Create a series star_ratings with the number of stars corresponding to each review in the dataset.
def star_rating():
    
    x = wine_catalog.points
    y = wine_catalog
    
    y.loc[x >= 95, 'points'] = 3
    
    y.loc[(x >= 85) & (x < 95), 'points'] = 2
    
    y.loc[(x < 80), 'points'] = 1
    
    return y
    
star_ratings = star_rating()
star_ratings.loc[star_ratings.points > 4]

# The first mistake I wrote
def __nonzero__(self):
    raise ValueError(f"The truth value of a {type(self).__name__} is ambiguous. "
                     "Use a.empty, a.bool(), a.item(), a.any() or a.all()."
                    )
    
__nonzero__(x)
