==> introduction_to_machine_learning/pandas_micro_course/04_grouping_and_sorting.ipynb <==
### Introduction
Maps allow us to transform data in a DataFrame or Series one value at a time for an entire column. However, often we want to gproup our data,

import pandas as pd

wine_store = pd.read_csv('../data/wine_catalog/wine_store_dataset.csv')

wine_store = wine_store.drop(columns=['Unnamed: 0'])

wine_store.head()

wine_store.sort_values(by='price')

### Groupwise analysis
One function we've been using heavily thus far is the `value_counts()` function.

The `groupby()` operations involves some combinations of splitting the object, applying a function, and combining the results.  
This can be used to group large amount of data and compute operations on these groups.

wine_store.groupby('points').points.count() # show us, how many times `points` it appeared.

`value_counts()` is just a shortcut to this `groupby()` operation.

For example, to get the expansive wine in each point value category, we can do the following:

wine_store.groupby('points').price.max()

This DataFrame is accessible to us directly using the `apply()` method, and we can manipulate the data in any way we see fit.

wine_store.groupby('winery').apply(lambda df: df.price.iloc[0])

wine_store.country.iloc[0]

wine_store.groupby(['country', 'province']).apply(lambda df: df.loc[df.points.idxmax()])
# x.to_csv('../data/wine_catalog/wine_store_country_winery.csv')

`agg()` lets you run a bunch of different functions on your DataFrame simultaneously.

x = wine_store.groupby(['country']).price.agg([len, min, max]).head()
x

The `groupby()` will allow you to do lots of really powerfull things with your dataset.

### Multi-indexes

In all of the examples we've seen thus far we've been working with DataFrame of Series objects with a single-label index.  
`groupby()` is slightly different in the fact that, depending on the operation we run we run, it will sometimes result in what is called a `multi-index`.

A `multi-index` differs from a regular index in that it has multiple levels.

countries_reviews = wine_store.groupby(['country', 'province']).points.agg([sum, max]) # sum of all price by province and country
countries_reviews

### agg()
```Python
DataFrame.agg(func=None, axis=0, *args, **kwargs)
```
Aggregate using one or more operations over the specified axis.

Multi-indeces have several methods for dealing with their tiered structure which are absent for single-level indices.

However, in general the multi-index method you will use most often is the one for converting back to a regular index, the `reset_index()` method:

countries_reviews.reset_index()

### Sorting

countries = countries_reviews.reset_index()
countries.sort_values(by=['max'])

### Exercise:

1) Create a `Series` whose count how many same prices? 

wine_store.groupby('price').size()

2) What is the best wine can I buy fro a given money?  
   Create `Series` whose index is wine prices and whose values is the maximum nuber of points a wine costing.

best_rating_per_price = wine_store.groupby('price').points.max()
best_rating_per_price

3) What are the min and max prices for each variety of wine.

price_extremes = wine_store.groupby('variety').price.agg(['min', 'max'])
price_extremes

4) What is the most cheap or most expensive wine varieties simultainously?

sort_varieties = price_extremes.sort_values(by=['min', 'max'], ascending=False, na_position='first')
sort_varieties

5) Province mean ratings.  
   Create `Series` whose index is province name and whose values is the average review score given out by points.

province_mean_ratings = wine_store.groupby('province').points.mean()
reviewer_mean_ratings 

6) Create a `Series` whose index is a `MultiIndex` of {'country', 'variety'} pairs.  
   Then sort the values in the `series` in descending order based on wine count.

country_variety_counts = wine_store.groupby(['country', 'variety']).size().sort_values(ascending=False, na_position='first')
country_variety_counts
