==> introduction_to_machine_learning/pandas_micro_course/05_data_types_and_missing_values.ipynb <==
### Introduction
In this tutorial, you'll learn how to investigate data types within a DataFrame or Series.  
You'll also learn how to find and replace entries.

import pandas as pd

wine_store = pd.read_csv('../data/wine_catalog/wine_store_dataset.csv')
wine_store = wine_store.drop(columns=['Unnamed: 0'])
wine_store.head()

### Dtypes
The data type for a column in a DataFrame or a Series is known as the `dtype` .   
You can use the `dtype` property to grab the type of specific column.

wine_store.price.dtype

Alternatively, the `dtypes` property return the `dtype` of every column in the DataFrame.

wine_store.dtypes

Data types tell us something about how pandas is storing the data internally.  
`float64` means that it's using a 64-bit floating point number;  
`int64` means a similarly sized integer instead, and so on.

`astype()` convert a column of one type into another type

wine_store['points'] = wine_store['points'].astype('int64')

wine_store.dtypes

wine_store.shape

# Replace all NaN elements with 0(zero's)
wine_store['price'] = wine_store['price'].fillna(0)

wine_store['price'] = wine_store['price'].astype('int64')

wine_store.dtypes

wine_store['price']

### Missing data
Entries missing values are given the value `NaN`, short for `Not a number` .  
For technical reasons these `NaN` values are always of the `float64` dtype.  

Pandas provides some methods specific to missing data.  
To select `NaN` entries you can use `pd.isnull()` or `pd.notnull()`

wine_store[pd.isnull(wine_store.country)]

wine_store[pd.isnull(wine_store.region_2)].head(2)

wine_store[pd.notnull(wine_store.country)].head()

# we can replace each NaN row with an`Unknown`.
wine_store['country'] = wine_store.country.fillna("Unknown")
wine_store.iloc[1123][0]

# we may have a non_null value that we  would like to replace.
wine_store['winery'] = wine_store.winery.replace("?", "Bodega Carmen Rodríguez")
wine_store.head()


wine_store.loc[wine_store.winery == 'Bodega Carmen Rodríguez']

### Exercise
What are the most common wine-product regions?

# We replace missing values "NaN" with "Unknown"
print(f'NaN = {pd.isnull(wine_store.region_2).sum()}')

wine_store['region_2'] = wine_store['region_2'].fillna('Unknown')

print(f'NaN = {pd.isnull(wine_store.region_2).sum()}')

# We create a `Series` with index names set from "region_2", and the values set from `count()` method.
# Then we sort the `Series` by descending. 
wine_store.groupby('region_2').region_2.count().sort_values(ascending=False)
