==> introduction_to_machine_learning/pandas_micro_course/06_renaming_and_combining.ipynb <==
### Introduction
Oftentimes data will come to us with column names, index names, or other naming convention that we are not satisfied with.
In that case, you'll learn how to use pandas functions to change the names of the offending entries to something better.

import pandas as pd

wine_store = pd.read_csv('../data/wine_catalog/wine_store_dataset.csv')
wine_store = wine_store.drop(columns='Unnamed: 0')
wine_store.head(2)

### Renaming
The function `rename()`, which lets you change index names and/or column names.

wine_store.rename(columns={'points': 'score'}).head(2)

`rename()` lets you rename index or column values by specifying a `index` or `column` keyword parameter, respectively.  
Here is an example using it to rename some elements of the index.

wine_store.rename(index={0: 'first', 1: 'second'}).head(2)

`set_index()` is usually more convenient for rename columns.  

Both the row index and the column index can have their `name` attribute.

wine_store.rename_axis('wines', axis='rows').rename_axis('fields', axis='columns').head(2)

### Combining
When performing operations on a dataset, we will sometimes need to combine different DataFrames and\or in non-trivial ways.  
Pandas has three core methods for doing this.  
In order of increasing complexity:
 - concat()
 - join()
 - merge()

canadian_youtube = pd.read_csv('../data/trending_youtube_video_stat/CAvideos/CAvideos.csv')
british_youtube = pd.read_csv('../data/trending_youtube_video_stat/GBvideos/GBvideos.csv')
russian_youtube = pd.read_csv('../data/trending_youtube_video_stat/RUvideos/RUvideos.csv', encoding='utf_8')

`concat()` this is usefull when we have data in different DataFrame or Series objects but having the same fields(columns).  
 If we want to study Youtube from multiple countries simultaneously, we can use concat() to smush them together:

pd.concat([canadian_youtube, british_youtube])

The middlemost combiner in terms of complexity is `join()` .  
`join()` lets you combine different DataFrame objects which have an index in common.

For example, to pull down videos that happened to be trending on the same day in both Canada and the UK, we could do the following:

left = canadian_youtube.set_index(['title', 'trending_date'])
right = british_youtube.set_index(['title', 'trending_date'])

left.join(right, lsuffix='_CAN', rsuffix='_UK')

The `lsuffix` and `rsuffix` parameters are necessary here because the data has the same column names in both British adn Canadian datasets.  
If this wasn't true (because, say, we'd renamed them beforehand) we wouldn't need them.

### Exercise


wine_store = pd.read_csv('../data/wine_catalog/wine_store_dataset.csv').drop(columns=['Unnamed: 0'])
wine_store.head(2)

1) Rename several columns: `region_1` and `region_2` with `region` and `locale`, respectively.

rename_columns = wine_store.rename(columns={'region_1': 'region', 'region_2': 'locale'})
rename_columns.head(2)

2) Set the index name in dataset to `wines`.

index_name = wine_store.rename_axis('wines', axis='columns')
index_name.head(2)

3) Create a `DataFrame` of products mentioned on *either* subreddit.

gaming = pd.read_csv('../data/06_renaming_and_combining_ex_03/gaming.csv')
gaming['subreddit'] = "r/gaming"

math = pd.read_csv('../data/06_renaming_and_combining_ex_03/math.csv')
math['subreddit'] = "r/math"

movie = pd.read_csv('../data/06_renaming_and_combining_ex_03/movies.csv')
movie['subreddit'] = "r/movies"

gaming

movie

combined_product

4) Using this, generate a dataset combining the two tables into one.

gaming.join(wine_store).head() # Join two table for fun)))
